requirements.txt
django
django-ratelimit
django-simple-captcha
# https://github.com/anymail/django-anymail
# https://github.com/fabiocaccamo/django-maintenance-mode
# https://github.com/django-haystack/django-haystack